Refactor for hwMapExt
