+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleOp/Constants.java
@@ -12,17 +12,6 @@ public class Constants {
         public static final String BACK_LEFT_MOTOR = "bl";
         public static final String BACK_RIGHT_MOTOR = "br";
 
-
-        public static final String LIFT_SERVO_A = "flickA";
-        public static final String LIFT_SERVO_B = "flickB";
-        public static final String LIFT_SERVO_C = "flickC";
-
-
-        public static final String INDEX_SENSOR_A = "indexSensorA";
-        public static final String INDEX_SENSOR_B = "indexSensorB";
-        public static final String INDEX_SENSOR_C = "indexSensorC";
-
-
         public static final DcMotorSimple.Direction FRONT_LEFT_DIRECTION = DcMotorSimple.Direction.REVERSE;
         public static final DcMotorSimple.Direction FRONT_RIGHT_DIRECTION = DcMotorSimple.Direction.FORWARD;
         public static final DcMotorSimple.Direction BACK_LEFT_DIRECTION = DcMotorSimple.Direction.REVERSE;
@@ -59,9 +48,23 @@ public class Constants {
 
 
 
+    }
+
+    public static class TransferConstants {
+        public static final String LIFT_SERVO_A = "flickA";
+        public static final String LIFT_SERVO_B = "flickB";
+        public static final String LIFT_SERVO_C = "flickC";
+
+
+        public static final String INDEX_SENSOR_A = "indexSensorA";
+        public static final String INDEX_SENSOR_B = "indexSensorB";
+        public static final String INDEX_SENSOR_C = "indexSensorC";
+
+
         // Lift Constants - TODO TUNE THESE PLS
         public static final double FLICK_POS_UP = 160; // with POS 0 at the top, with cw rotation
         public static final double FLICK_POS_DOWN = 30; // with POS 0 at the top, with cw rotation
+
     }
 
     // Controller Constants
