+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/teleOp/finalTeleOp.java
@@ -3,7 +3,7 @@ package org.firstinspires.ftc.teamcode.teleOp;
 import com.qualcomm.robotcore.eventloop.opmode.LinearOpMode;
 import com.qualcomm.robotcore.eventloop.opmode.TeleOp;
 
-import org.firstinspires.ftc.teamcode.Hware.hwMap;
+import org.firstinspires.ftc.teamcode.Hware.hwMapExt;
 import org.firstinspires.ftc.teamcode.subsystems.DriveTrain;
 import org.firstinspires.ftc.teamcode.subsystems.GameState;
 import org.firstinspires.ftc.teamcode.subsystems.RobotState;
@@ -15,7 +15,7 @@ public class finalTeleOp extends LinearOpMode {
 
     @Override
     public void runOpMode() throws InterruptedException {
-        hwMap hardware = new hwMap(hardwareMap);
+        hwMapExt hardware = new hwMapExt(hardwareMap);
         stateMachine = new StateMachine(hardware);
 
         waitForStart();
